Updated packages, migration to MUI v5, src restructuring
