+++ src/components/templates/Summary.js
@@ -1,5 +1,5 @@
 import React from "react";
-import {Typography} from "@material-ui/core";
+import {Typography} from "@mui/material";
 
 export default function Summary({sections}) {
     return (
