+++ src/styles/app.styles.js
@@ -1,4 +1,4 @@
-import {makeStyles} from "@material-ui/core/styles";
+import {makeStyles} from "@mui/styles";
 
 export const styles = makeStyles(theme => ({
     appBar: {
